0.3.0 Release
This commit signifies the 0.3.0 release.  This release contains breaking changes

This release modifies the create and edit nodes to look for a data property on the payload. Either node will use this property to set FileMaker field or portal data.

If the data object does not contain a fieldData property it will assume all properties except portalData should be merged into a created fieldData property.

This release also changes the behavior of the utility nodes transform, recordId, and fieldData. The utility nodes will now all look for a data property on the payload.

Signed-off-by: Lui de la Parra <lui@mutesymphony.com>