0.5.0 Release
This commit signifies the 0.5.0 Release. This release contains breaking changes.

The upload node now looks for a file property on the message payload instead of filename. The upload node can accept either a string or a buffer.

The globals node now looks for a data property on the payload.

This project just hit 100% test coverage!!

Signed-off-by: Lui de la Parra <lui@mutesymphony.com>