0.9.8 Release
This commit signifies the 0.9.8 release. This release updates node documentation and revises the container data node to ensure the destination will exist.

This release contains breaking changes. The client node namspacing has been brought inline with project styles and the application and server configuration properties are now moved to a credentials object.

This was done to protect sensitive connection information to be exposed through sharing flows.

Signed-off-by: Lui de la Parra <Lui@mutesymphony.com>