0.9.9 Release
This commit signifies the 0.9.9 release. This release adds the ability to provide a custom name for each node. This release contains an update to each node’s editor panel adding icons.

This release also updates the upload node to upload files to either a specific record or a create a new record.

Finally this release contains minor bugfixes to the create and edit nodes.

Signed-off-by: Lui de la Parra <lui@mutesymphony.com>