2.2.0 Release
This commit signifies the 2.2.0 release. This release adds the ability to persist the Data API client to a variety of datastores. By default the client will continue to use an in-memory store.
The client can now also be configured via the Node-RED settings file to use filesystem or MongoDB storage as well.

Additionally this release adds a status node and modifies the logout node. The status node will return the client’s current data usage, pending and queued requests, and open Data API sessions. The logout node has been modified to also accept a session id as the session to logout.